requirements.txt
appdirs==1.4.0
Appium-Python-Client==0.22
APScheduler==3.3.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.5.1
cached-property==1.3.0
cffi==1.9.1
colorama==0.3.7
cryptography==1.6
cx-Oracle==5.2.1
cycler==0.10.0
decorator==4.0.11
defusedxml==0.4.1
docutils==0.13.1
enum34==1.1.6
eyes-selenium==3.2
fake-factory==0.7.2
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
Gecko==1.0.19
idna==2.1
ipaddress==1.0.17
ipython==5.1.0
ipython-genutils==0.1.0
isodate==0.5.4
jenkinsapi==0.3.3
jsonpatch==1.14
jsonpointer==1.10
lxml==3.6.4
matplotlib==2.0.0
mysql-connector-python==2.1.4
natsort==5.0.2
numpy==1.12.0
packaging==16.8
paramiko==2.0.2
pathlib2==2.1.0
pickleshare==0.7.4
prompt-toolkit==1.0.8
pyasn1==0.1.9
pycparser==2.17
Pygments==2.1.3
pyparsing==2.1.10
pypng==0.0.18
pyserial==3.2.1
python-dateutil==2.5.3
pytz==2016.7
PyYAML==3.12
requests==2.11.1
rinse==0.4.0
Robot-AppEyes==1.2
robotframework==3.0.1
robotframework-angularjs==0.0.4
robotframework-excellibrary==0.0.2
robotframework-extendedselenium2library==0.9.1
robotframework-faker==3.0.0
robotframework-httplibrary==0.4.2
robotframework-requests==0.4.5
robotframework-ride==1.5.2.1
robotframework-selenium2library==1.8.0
robotframework-sshlibrary==2.1.2
robotframework-Web2Library==0.1.327.dev0
robotframework-YamlVariablesLibrary==0.0.5.dev0
selenium==3.0.2
simplegeneric==0.8.1
six==1.10.0
suds==0.4
traitlets==4.3.1
tzlocal==1.3
waitress==1.0.0
wcwidth==0.1.7
WebOb==1.6.1
WebTest==2.0.23
win-unicode-console==0.5
wrapt==1.10.8
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.2.0
zeep==0.22.0
